Isolated the test to it's own JVM instance. So it doesn't pollute the environment of other tests.
